Author classes used for data manipulation
Create the formal interfaces of the `DataIO`, `DataObject`,
`DataIOStrategy` classes, and provide and empty implementation in
`AndroidDataIO`. Mark `FileIO` and related as deprecated.

Our Android apps require some method of interacting with files on disk
to store -- at a minimum -- configuration settings. The previous
`FileIO` interface does not provide all desired features.

Several parts of the documentation are incomplete are require
completion.