Modify behaviour when accessing location manager
Explicitly catch the `java.lang.SecurityException` that would be thrown
in the event the app is denied access to the devices location manager,
and dump information about the issue to our logging system.

Our application might not necessary have access to the location manager
that is present on the device. Under the current behaviour our framework
places the application in an undefined state.

The use of `checkSelfPermission(Context, String)` was avoided because
our lowest target API is below level 23 (Marshmallow v6.0.0).