Catch failures when calling `setMyLocationEnabled`
Dump an error message to our logging system in the event we are unable
to access the user's current location.

We are not guaranteed to have access to any location data from the user.
If we do not attempt to catch events we risk our applications falling
into a undeterminable state.

Also update the class `Graphics` authorship information in it's Javadoc
header (adding both the original author, and those that made minor
changes).