Incrememnt framework version details
Shift the value of `version_code` to `3`, and the value of
`version_name` to `v1.0.0`.

Version tracking information needs to be stored in the framework so that
applications using it can be aware of what they are using.

This release marks the final release of Mother. Bug fixing is only done
from this point out.