Shift version details into the manifest
Remove the string resources 'version_code', and 'version_name', and
place the values directly in there targets inside the mainfest file.

The `android:versionCode` and `android:versionName` attributes of the
`manifest` tag are not allowed to be dynamic resources.

Because 'strings.xml' contained only these resources, the file is
removed.